LayoutTests/accessibility/mac/aria-popup-expected.txt

Example
This tests that when an ARIA button has aria-haspopup=true, it will be a popup button

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS button.role is 'AXRole: AXPopUpButton'
PASS button.isPressActionSupported() is true
PASS successfullyParsed is true

TEST COMPLETE
